Add modification check before starting webserver 	(only start webserver if script is not modified)

+++ autoshutdown.sh
@@ -98,6 +98,17 @@ MY_UUID=$(uuidgen)
 # ########################################################
 
 # -------------------------------------------
+
+check_hash_script_modified(){
+    MD5_HASHSCRIPT_SAVED=$(cat $HASHSCRIPTFILE)
+    if [ "$MD5_HASHSCRIPT_SAVED" != "$MD5_HASHSCRIPT" ]; then
+		RET=1
+    else
+        RET=0
+    fi
+	echo $RET
+}
+
 #######################################
 # Starts an asynchron webserver
 # Globals:
@@ -108,42 +119,47 @@ MY_UUID=$(uuidgen)
 #   
 #######################################
 init_webserver_shutdown(){
-	
-	# shutdown all previous instances
-	WEBSERVER_INSTANCES=$(ps -ef|grep -v grep|grep $WEBSERVER_SHUTDOWN_SCRIPT|wc -l)
-	if [ $WEBSERVER_INSTANCES -ne 0 ]; then
-		writelog "I" "Kill all instances ($WEBSERVER_INSTANCES) of '$WEBSERVER_SHUTDOWN_SCRIPT'"
-		pkill -f $WEBSERVER_SHUTDOWN_SCRIPT >/dev/null 2>&1
-		sleep 5
-	fi
-
-	WEBSERVER_INSTANCES=$(ps -ef|grep -v grep|grep $WEBSERVER_SHUTDOWN_SCRIPT|wc -l)
-	WEBSERVER_STARTUP_FAILED=0
-	WEBSERVER_STARTUP_COUNTER=0
-	writelog "I" "Waiting for webserver to start (instances=$WEBSERVER_INSTANCES; loop=$WEBSERVER_STARTUP_COUNTER)"
-	while [[ $WEBSERVER_STARTUP_FAILED -eq 0 ]] && [[ $WEBSERVER_INSTANCES -eq 0 ]]; do
-		WEBSERVER_STARTUP_COUNTER=$((WEBSERVER_STARTUP_COUNTER+1))
-
-		if [ $WEBSERVER_INSTANCES -eq 0 ]; then 
-			python "$WEBSERVER_SHUTDOWN_SCRIPT" --port $WEBSERVER_SHUTDOWN_PORT --uuid $MY_UUID --spath $WEBSERVER_SHUTDOWN_URL --tpath $WEBSERVER_TEST_URL &
-		fi
-		
-		sleep 2
+	# check if script file is modified? if true, the webserver would be restarted after reloading/restarting the script
+	SCRIPT_MODIFIED=$(check_hash_script_modified)
+	if [ $SCRIPT_MODIFIED -eq 0 ]; then
+		# shutdown all previous instances
 		WEBSERVER_INSTANCES=$(ps -ef|grep -v grep|grep $WEBSERVER_SHUTDOWN_SCRIPT|wc -l)
-		writelog "I" "Waiting for webserver to start (instances=$WEBSERVER_INSTANCES; loop=$WEBSERVER_STARTUP_COUNTER)"
-
-		if [ $WEBSERVER_STARTUP_COUNTER -gt 5 ]; then
-			WEBSERVER_STARTUP_FAILED=1
+		if [ $WEBSERVER_INSTANCES -ne 0 ]; then
+			writelog "I" "Kill all instances ($WEBSERVER_INSTANCES) of '$WEBSERVER_SHUTDOWN_SCRIPT'"
+			pkill -f $WEBSERVER_SHUTDOWN_SCRIPT >/dev/null 2>&1
+			sleep 5
 		fi
-	done
 
-	if [ $WEBSERVER_INSTANCES -eq 0 ]; then
-		writelog "E" "Failed to start webserver!"
+		WEBSERVER_INSTANCES=$(ps -ef|grep -v grep|grep $WEBSERVER_SHUTDOWN_SCRIPT|wc -l)
+		WEBSERVER_STARTUP_FAILED=0
+		WEBSERVER_STARTUP_COUNTER=0
+		writelog "I" "Waiting for webserver to start (instances=$WEBSERVER_INSTANCES; loop=$WEBSERVER_STARTUP_COUNTER)"
+		while [[ $WEBSERVER_STARTUP_FAILED -eq 0 ]] && [[ $WEBSERVER_INSTANCES -eq 0 ]]; do
+			WEBSERVER_STARTUP_COUNTER=$((WEBSERVER_STARTUP_COUNTER+1))
+
+			if [ $WEBSERVER_INSTANCES -eq 0 ]; then 
+				python "$WEBSERVER_SHUTDOWN_SCRIPT" --port $WEBSERVER_SHUTDOWN_PORT --uuid $MY_UUID --spath $WEBSERVER_SHUTDOWN_URL --tpath $WEBSERVER_TEST_URL &
+			fi
+			
+			sleep 2
+			WEBSERVER_INSTANCES=$(ps -ef|grep -v grep|grep $WEBSERVER_SHUTDOWN_SCRIPT|wc -l)
+			writelog "I" "Waiting for webserver to start (instances=$WEBSERVER_INSTANCES; loop=$WEBSERVER_STARTUP_COUNTER)"
+
+			if [ $WEBSERVER_STARTUP_COUNTER -gt 5 ]; then
+				WEBSERVER_STARTUP_FAILED=1
+			fi
+		done
+
+		if [ $WEBSERVER_INSTANCES -eq 0 ]; then
+			writelog "E" "Failed to start webserver!"
+		else
+			writelog "I" "Shutdown webserver (external call) set to 'http://$MY_PUBLIC_IP:$WEBSERVER_SHUTDOWN_PORT_EXTERNAL/$MY_UUID/$WEBSERVER_SHUTDOWN_URL'"
+			writelog "I" "Shutdown webserver (local call) set to 'http://localhost:$WEBSERVER_SHUTDOWN_PORT/$MY_UUID/$WEBSERVER_TEST_URL'"
+			notification "$MYNAME" "$MESSAGE_WEBSERVER_SHUTDOWN_START"
+			notification "$MYNAME" "http://$MY_PUBLIC_IP:$WEBSERVER_SHUTDOWN_PORT_EXTERNAL/$MY_UUID/$WEBSERVER_SHUTDOWN_URL"
+		fi
 	else
-		writelog "I" "Shutdown webserver (external call) set to 'http://$MY_PUBLIC_IP:$WEBSERVER_SHUTDOWN_PORT_EXTERNAL/$MY_UUID/$WEBSERVER_SHUTDOWN_URL'"
-		writelog "I" "Shutdown webserver (local call) set to 'http://localhost:$WEBSERVER_SHUTDOWN_PORT/$MY_UUID/$WEBSERVER_TEST_URL'"
-		notification "$MYNAME" "$MESSAGE_WEBSERVER_SHUTDOWN_START"
-		notification "$MYNAME" "http://$MY_PUBLIC_IP:$WEBSERVER_SHUTDOWN_PORT_EXTERNAL/$MY_UUID/$WEBSERVER_SHUTDOWN_URL"
+		writelog "I" "Script modified, do not start webserver in this instance... Please wait for reloading."
 	fi
 }
